# openapi.yml
openapi: "3.1.0"
info:
  title: Clawkson API
  description: API for the Clawkson AI assistant platform
  version: 0.1.0
  license:
    name: MIT
servers:
  - url: http://localhost:47821
    description: Local development server
paths:
...
  /api/agents/{id}/container:
    get:
      summary: Get container status for an agent
      operationId: getContainerStatus
      tags: [Containers]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Container status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContainerStatus"
        "404":
          description: No container for this agent
    delete:
      summary: Remove container and workspace
      operationId: removeContainer
      tags: [Containers]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204":
          description: Removed
        "503":
          description: Docker not available
  /api/agents/{id}/container/start:
    post:
      summary: Create and start a container for an agent
      operationId: startContainer
      tags: [Containers]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Container started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContainerStatus"
        "400":
          description: Container not enabled for agent
  /api/agents/{id}/container/stop:
    post:
      summary: Stop a running container
      operationId: stopContainer
      tags: [Containers]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204":
          description: Stopped
  /api/agents/{id}/container/logs:
    get:
      summary: Get container logs
      operationId: getContainerLogs
      tags: [Containers]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: tail, in: query, schema: { type: integer, default: 100 } }
      responses:
        "200":
          description: Container logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: string
  /api/agents/{id}/container/exec:
    post:
      summary: Execute a command in the container
      operationId: execInContainer
      tags: [Containers]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecRequest"
      responses:
        "200":
          description: Execution result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecResult"
...
  /api/connectors:
    get:
      summary: List all connectors
      operationId: listConnectors
      tags: [Connectors]
      responses:
        "200":
          description: List of connectors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Connector"
    post:
      summary: Create a connector
      operationId: createConnector
      tags: [Connectors]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConnectorRequest"
      responses:
        "200":
          description: The created connector
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connector"
  /api/connectors/{id}:
    get:
      summary: Get a connector by ID
      operationId: getConnector
      tags: [Connectors]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The connector
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connector"
        "404":
          description: Not found
    patch:
      summary: Update a connector (e.g. toggle enabled)
      operationId: patchConnector
      tags: [Connectors]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchConnectorRequest"
      responses:
        "200":
          description: The updated connector
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connector"
        "404":
          description: Not found
    delete:
      summary: Delete a connector
      operationId: deleteConnector
      tags: [Connectors]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Knowledge Base ──────────────────────────────────────────
...
  /api/llm-connectors:
    get:
      summary: List all LLM connectors (API keys masked)
      operationId: listLlmConnectors
      tags: [LLM Connectors]
      responses:
        "200":
          description: List of connectors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LlmConnector"
    post:
      summary: Create an LLM connector
      operationId: createLlmConnector
      tags: [LLM Connectors]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLlmConnectorRequest"
      responses:
        "200":
          description: Created connector (API key masked)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LlmConnector"
  /api/llm-connectors/test:
    post:
      summary: Test an LLM connector configuration without saving it
      operationId: testLlmConnector
      tags: [LLM Connectors]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLlmConnectorRequest"
      responses:
        "200":
          description: Test result with latency
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestConnectionResponse"
  /api/llm-connectors/{id}:
    get:
      summary: Get an LLM connector
      operationId: getLlmConnector
      tags: [LLM Connectors]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: The connector
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LlmConnector"
        "404":
          description: Not found
    patch:
      summary: Update an LLM connector
      operationId: patchLlmConnector
      tags: [LLM Connectors]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchLlmConnectorRequest"
      responses:
        "200":
          description: Updated connector
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LlmConnector"
        "404":
          description: Not found
    delete:
      summary: Delete an LLM connector
      operationId: deleteLlmConnector
      tags: [LLM Connectors]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Uploads (File Attachments) ────────────────────────────────────
...
components:
  schemas:
...
    ContainerStatus:
      type: object
      required: [agent_id, state, image, workspace_path]
      properties:
        agent_id:
          type: string
          format: uuid
        state:
          type: string
          enum: [creating, running, stopped, removing]
        image:
          type: string
        workspace_path:
          type: string
    ExecRequest:
      type: object
      required: [command]
      properties:
        command:
          type: string
          description: Shell command to execute
        timeout:
          type: integer
          description: Timeout in seconds (default 30, max 300)
          nullable: true
        output_dir:
          type: string
          nullable: true
          description: >
            Workspace-relative directory to scan for output files after execution
            (default "outputs"). Set to empty string to disable output collection.
    ExecResult:
      type: object
      required: [stdout, stderr, exit_code, timed_out]
      properties:
        stdout:
          type: string
        stderr:
          type: string
        exit_code:
          type: integer
        timed_out:
          type: boolean
        output_files:
          type: array
          nullable: true
          description: Files found in the output directory after execution
          items:
            $ref: "#/components/schemas/OutputFile"
    OutputFile:
      type: object
      required: [path, size]
      properties:
        path:
          type: string
          description: Path relative to the workspace root
        size:
          type: integer
          description: File size in bytes
...
    Connector:
      type: object
      required: [id, user_id, name, connector_type, enabled, config, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
          description: The user who owns this connector
        name:
          type: string
        connector_type:
          $ref: "#/components/schemas/ConnectorType"
        enabled:
          type: boolean
        config:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ConnectorType:
      type: string
      enum: [telegram, gmail, slack, azure_devops, custom]
    CreateConnectorRequest:
      type: object
      required: [name, connector_type, config]
      properties:
        name:
          type: string
        connector_type:
          $ref: "#/components/schemas/ConnectorType"
        config:
          type: object
    PatchConnectorRequest:
      type: object
      properties:
        enabled:
          type: boolean
...
    LlmProviderType:
      type: string
      enum: [azure, open_router, open_ai, custom]
      description: The LLM provider backend type
    LlmConnector:
      type: object
      required: [id, name, provider_type, model, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        provider_type:
          $ref: "#/components/schemas/LlmProviderType"
        model:
          type: string
        api_base_url:
          type: string
          nullable: true
          description: Required for custom provider; optional for others
        azure_deployment:
          type: string
          nullable: true
          description: Azure OpenAI deployment name
        azure_api_version:
          type: string
          nullable: true
          description: Azure OpenAI API version (e.g. 2024-02-01)
        api_key_masked:
          type: string
          nullable: true
          description: Masked API key (e.g. sk-or-••••1234); never the real key
        created_at:
          type: string
          format: date-time
    CreateLlmConnectorRequest:
      type: object
      required: [name, provider_type, model, api_key]
      properties:
        name:
          type: string
        provider_type:
          $ref: "#/components/schemas/LlmProviderType"
        model:
          type: string
        api_key:
          type: string
          description: Plain-text API key (stored encrypted in the database; masked on all responses)
        api_base_url:
          type: string
          nullable: true
        azure_deployment:
          type: string
          nullable: true
        azure_api_version:
          type: string
          nullable: true
    PatchLlmConnectorRequest:
      type: object
      properties:
        name:
          type: string
        provider_type:
          $ref: "#/components/schemas/LlmProviderType"
        model:
          type: string
        api_key:
          type: string
          nullable: true
        api_base_url:
          type: string
          nullable: true
        azure_deployment:
          type: string
          nullable: true
        azure_api_version:
          type: string
          nullable: true
...
    TestConnectionResponse:
      type: object
      required: [ok, latency_ms]
      properties:
        ok:
          type: boolean
        latency_ms:
          type: integer
          format: int64
        error:
          type: string
          nullable: true

    # ── Settings ───────────────────────────────────────────────
...